the interface needs to be changed as well..
git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@31411 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ channels/misdn/isdn_lib.h
@@ -350,7 +350,7 @@ void manager_bchannel_deactivate(struct misdn_bchannel * bc);
 
 int misdn_lib_tx2misdn_frm(struct misdn_bchannel *bc, void *data, int len);
 
-void manager_ph_control(struct misdn_bchannel *bc, long c1, long c2);
+void manager_ph_control(struct misdn_bchannel *bc, int c1, int c2);
 
 
 int misdn_lib_port_restart(int port);
